Refactor most of GhciSession into GhciEvaluator
GhciSession is going to become a bit fatter. This hides a GhciSession
implementation detail behind a package-private class.

It does break a test.
